move weights.rs to src/test
